+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # brote
 
+## 1.0.7
+
+### Patch Changes
+
+- adding import/export logic for the first version
+
 ## 1.0.6
 
 ### Patch Changes