+++ src/components/pageMeal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # page-meal
 
+## 0.0.5
+
+### Patch Changes
+
+- adding import/export logic for the first version
+
 ## 0.0.4
 
 ### Patch Changes