+++ src/components/pageOpenCal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # page-opencal
 
+## 0.0.4
+
+### Patch Changes
+
+- adding import/export logic for the first version
+
 ## 0.0.3
 
 ### Patch Changes